+++ packages/components/src/auto-grid/index.ts
@@ -1,9 +1,13 @@
-import type { ExtractPropTypes } from 'vue';
-import AutoGrid, { autoGridProps } from './src/index';
+import AutoGrid from './src/index.vue';
+import type { AutoGridProps } from './src/types';
+
+type AutoGridInstance = InstanceType<typeof AutoGrid>;
+
+export type {
+  AutoGridInstance as MixteAutoGridInstance,
+  AutoGridProps as MixteAutoGridProps,
+};
 
 export {
   AutoGrid as MixteAutoGrid,
-  autoGridProps as mixteAutoGridProps,
 };
-
-export type MixteAutoGridProps = ExtractPropTypes<typeof autoGridProps>;